添加，根据id查询，dao，xml错误。

+++ gymnasium_common/src/main/java/com/vky/pojo/Field.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class Field {
-    private Long fieldNo;
+    private Integer fieldNo;
     private String fieldName;
     private String fieldType;
     private Integer fieldCost;
